按照lilith的要求，所生成的luc文件严格是32位的。

+++ src/ldump.c
@@ -54,10 +54,10 @@ static void DumpVector(const void* b, int n, size_t size, DumpState* D) {
 
 static void DumpString(const TString* s, DumpState* D) {
   if (s==NULL || getstr(s)==NULL) {
-    size_t size=0;
+    int size=0;
     DumpVar(size,D);
   } else {
-    size_t size=s->tsv.len+1;		/* include trailing '\0' */
+    int size=s->tsv.len+1;		/* 包含末尾的'\0'字符 */
     DumpVar(size,D);
     DumpBlock(getstr(s),size,D);
   }
