+++ src/nundump.c
@@ -82,7 +82,8 @@ static lua_Number LoadNumber(LoadState* S) {
 
 /* 从字节码文件中读取字符串 */
 static TString* LoadString(LoadState* S) {
-  size_t size;
+  //size_t size;
+  int size;
   global_State* g=G(S->L);
   LoadVar(S,size);                        /* 读取字符串长度 */
   if (size==0)
